+++ src/components/carousel/Carousel.js
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from 'react';
+import React from 'react';
 import { useSelector, useDispatch } from 'react-redux';
 import Carousel from 'react-multi-carousel';
 import 'react-multi-carousel/lib/styles.css';
@@ -12,9 +12,7 @@ import {
   VStack,
   Button,
   Text,
-  Image,
   Divider,
-  Spinner,
 } from '@chakra-ui/react';
 import { formatPrices } from '../../utils/product_utils/product_utils';
 import { useContext } from 'react';
@@ -23,52 +21,8 @@ import { addItemToCart } from '../../utils/cart_utils/cartUtils';
 import { addItemToCartAction } from '../../redux/cart/cartActions';
 import AlertMessage from '../ui/AlertMessage.jsx';
 
-const responsive = {
-  superLargeDesktop: {
-    // the naming can be any, depends on you.
-    breakpoint: { max: 4000, min: 3000 },
-    items: 5,
-  },
-  desktop: {
-    breakpoint: { max: 3000, min: 921 },
-    items: 3,
-  },
-  tablet: {
-    breakpoint: { max: 920, min: 520 },
-    items: 2,
-  },
-  mobile: {
-    breakpoint: { max: 464, min: 0 },
-    items: 1,
-  },
-};
-
-const CustomImage = ({ ...props }) => {
-  const [url, setUrl] = useState('');
-
-  useEffect(() => {
-    const fetchImage = async (path) => {
-      const fullPath =
-        /* 'http://127.0.0.1:5001/costofinal-b391b/us-central1/images/api/images'; */
-        'https://us-central1-costofinal-b391b.cloudfunctions.net/images/api/images';
-
-      const fetchResponse = await fetch(fullPath, {
-        method: 'POST',
-        body: path,
-      });
-
-      const fetchData = await fetchResponse.json();
-
-      console.log('fetchData', fetchData);
-
-      return fetchData.url;
-    };
-
-    fetchImage(props.src).then((url) => setUrl(url));
-  }, [setUrl, url]);
-
-  return url ? <Image {...props} src={url} /> : <Spinner />;
-};
+import responsive from './responsive';
+import CustomImage from './CustomImage';
 
 const CustomCarousel = ({ items }) => {
   const onOpenLogin = useContext(OpenLoginContext);
