+++ src/components/carousel/ChakraBox.js
@@ -1,0 +1,9 @@
+import { shouldForwardProp, chakra } from '@chakra-ui/react';
+import { motion, isValidMotionProp } from 'framer-motion';
+
+const ChakraBox = chakra(motion.div, {
+  shouldForwardProp: (prop) =>
+    isValidMotionProp(prop) || shouldForwardProp(prop),
+});
+
+export default ChakraBox;
