+++ src/components/carousel/Limpieza.js
@@ -1,16 +1,11 @@
 import React from 'react';
-
-// import Carousel from 'react-multi-carousel';
-// import 'react-multi-carousel/lib/styles.css';
-import './style.css';
-import { Box, Card, CardBody, CardHeader, Text } from '@chakra-ui/react';
-import { NavLink } from 'react-router-dom';
+import { Card, CardBody, CardHeader, Box, Text } from '@chakra-ui/react';
 import CustomImage from './CustomImage';
 import ChakraBox from './ChakraBox';
 
-const NewsCarousel = () => {
-  /* const [isStopped, setisStopped] = useState(true); */
+import { NavLink } from 'react-router-dom';
 
+const Limpieza = () => {
   return (
     <Box
       zIndex={'500'}
@@ -40,24 +35,23 @@ const NewsCarousel = () => {
         marginBottom="30px">
         <CustomImage src={'/new.png'} width="100%" height={'100%'} />
       </ChakraBox>
-
-      <NavLink to={'products/art-de-pesca'}>
+      <NavLink to={'products/limpieza'}>
         <Card size={{ base: 'md', sm: 'sm' }} align={'center'} justify="center">
           <CardHeader textAlign={'center'}>
-            {/* <VStack alignContent={'center'}> */}
-            <Text fontSize={{ base: 'lg', sm: 'sm' }}>Artículos de Pesca</Text>
+            <Text fontSize={{ base: 'lg', sm: 'sm' }}>
+              Artículos de Limpieza
+            </Text>
           </CardHeader>
           <CardBody>
             <CustomImage
-              src={'/cards/multifilamentos.jpg'}
+              src={'/cards/limpieza.jpg'}
               /* h={{ base: '9rem', sm: '4rem' }} */
             />
           </CardBody>
-          {/* </VStack> */}
         </Card>
       </NavLink>
     </Box>
   );
 };
 
-export default NewsCarousel;
+export default Limpieza;
