+++ src/components/carousel/responsive.js
@@ -1,0 +1,21 @@
+const responsive = {
+  superLargeDesktop: {
+    // the naming can be any, depends on you.
+    breakpoint: { max: 4000, min: 3000 },
+    items: 5,
+  },
+  desktop: {
+    breakpoint: { max: 3000, min: 921 },
+    items: 3,
+  },
+  tablet: {
+    breakpoint: { max: 920, min: 520 },
+    items: 2,
+  },
+  mobile: {
+    breakpoint: { max: 464, min: 0 },
+    items: 1,
+  },
+};
+
+export default responsive;
