+++ src/data/itemsToCarrousel.json
@@ -25,9 +25,9 @@
     "image": "/assets/caroussel/polipropileno_codo_90_hh.jpg"
   },
   {
-    "id": "1b30f2cd-fb08-4179-9719-f2269eea557c",
-    "codigo": "R2042",
-    "image": "/assets/caroussel/rejillas_ventilacion_20x40_800.jpg"
+    "id": "1b9f5fb7-16b6-41e8-aedc-638d347c0072",
+    "codigo": "R20AI",
+    "image": "/assets/caroussel/ryt_rejas_piso.jpg"
   },
   {
     "id": "270cfdbb-37a2-46d4-a56a-cc04a1020bfb",
